requirements.txt
fastapi
uvicorn
requests~=2.32.4
ollama
python-dotenv~=1.1.1
langchain~=0.3.26
langchain-core~=0.3.68
langchain-ollama
langchain-chroma
chromadb~=1.0.15
h11~=0.16.0
pip~=25.0.1
build~=1.2.2.post1
wheel~=0.45.1
protobuf~=5.29.5
typing_extensions~=4.14.1
sniffio~=1.3.1
click~=8.2.1
httpcore~=1.0.9
Pygments~=2.19.2
rich~=14.0.0
idna~=3.10
certifi~=2025.7.9
httpx~=0.28.1
langmem~=0.0.28
pydantic~=2.11.7
anyio~=4.9.0
pydantic_core~=2.33.2
annotated-types~=0.7.0
setuptools~=78.1.0
PyYAML~=6.0.2
multidict~=6.6.3
propcache~=0.3.2
typing_extensions~=4.14.1
attrs~=25.3.0
numpy~=2.3.1
charset-normalizer~=3.4.2
overrides~=7.7.0
orjson~=3.10.18
xxhash~=3.5.0
yarl~=1.20.1
aiohttp~=3.12.14
aiosignal~=1.4.0
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
uvloop~=0.21.0
urllib3~=2.5.0
pydantic_core~=2.33.2
tenacity~=9.1.2
langchain-community~=0.3.27
langsmith~=0.4.5
openai~=1.95.0
langchain-text-splitters~=0.3.8
SQLAlchemy~=2.0.41
ormsgpack~=1.10.0
anthropic~=0.57.1
zstandard~=0.23.0
requests-toolbelt~=1.0.0
packaging~=24.2
importlib_metadata~=8.7.0
marshmallow~=3.26.1
jsonpointer~=3.0.0
jsonpatch~=1.33
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
pydantic-settings~=2.10.1
mypy_extensions~=1.1.0
httpx-sse~=0.4.1
tiktoken~=0.9.0
langchain-google-genai~=2.1.7
typing_extensions~=4.14.1
paddleocr~=3.1.0
typing_extensions~=4.14.1
pytesseract~=0.3.13
typing_extensions~=4.14.1
pillow~=11.3.0
typing_extensions~=4.14.1
PyMuPDF~=1.26.3
typing_extensions~=4.14.1
gradio~=5.38.2
typing_extensions~=4.14.1
langgraph~=0.6.0
typing_extensions~=4.14.1